# demos/01-dictionary/backend/shipping-service/docs/swagger.yaml
basePath: /
definitions:
...
  rest.errorResponse:
    properties:
      error:
        type: string
    type: object
...
  rest.switchTenantRequest:
    properties:
      tenant:
        type: string
    type: object
  rest.tenantResponse:
    properties:
      available:
        items:
          type: string
        type: array
      tenant:
        type: string
    type: object
host: localhost:7200
info:
  contact: {}
  description: Shipping domain backend for the NATS Tech Lab POC. Demonstrates JetStream
    event sourcing, NATS KV projections (Shape A / Shape B), and pure event reconstruction
    (Shape C).
  title: EventSourcing CQRS POC — Shipping API
  version: "1.0"
paths:
...
  /api/tenant/switch:
    post:
      consumes:
      - application/json
      description: Reconnects shipping-service's tenant-scoped NATS connection under
        a different account and rebinds every ship/container resource to it. The previous
        tenant's data becomes unreachable because the server enforces the account
        boundary, not because of an application-level filter.
      parameters:
      - description: target tenant
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/rest.switchTenantRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.tenantResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.errorResponse'
      summary: Switch the active tenant (Phase 13b)
      tags:
      - tenant
...
swagger: "2.0"